migrations actually work now

+++ migrations/2018-03-14-005302_origin_members/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS origin_members (
-  origin_id int REFERENCES origins(id),
+  origin_id bigint REFERENCES origins(id),
   origin_name text,
-  account_id int,
+  account_id bigint,
   account_name text,
   created_at timestamptz DEFAULT now(),
   updated_at timestamptz DEFAULT now(),
